Remove bundled libjpeg and use system lib instead
In the bundled copy of libjpeg, jpeg_stdio_src was (confusingly)
modified to read a buffer not a FILE, as well as a custom
jpeg_memory_src. ROQ used the stdio version, the renderer the
memory one.
Only keept the memory version, modified ROQ to use that - though
it takes a maximum size the stdio version didn't, so used the filesize.

Had to include the jpegint.h of libjpeg-turbo-1.1.90 for roq's internal
use