Fixed the Free16 function in Heap.cpp to take into account the size of pointers.
Allocate16 already too this into account, but Free16 didn't. However, I updated Allocate16 to use sizeof(std::ptrdiff_t) instead of sizeof(size_t) to be more correct. (size_t isn't guaranteed to be the size of a pointer, just the max size of an array on the current platform.)